ref(modules): move eval-result VOs to Phel\Shared\Eval
EvalResult, EvalError and StackFrame moved from Run\Domain\Repl to the
leaf Phel\Shared\Eval namespace, so RunFacadeInterface::structuredEval()
and its Nrepl/Watch consumers no longer reach into another module's
Domain (mirrors the earlier NamespaceInformation move).

The compiler orchestration that produced the result (formerly the
EvalResult::fromEval factory, which depended on CompilerFacade and the
GlobalEnvironment) now lives in Run\Application\StructuredEvaluator,
keeping the moved VOs logic-free as the Shared leaf requires. The
fromEval test cases moved to StructuredEvaluatorTest; the pure-VO cases
to a Shared EvalResultTest.

Related to #2378
