+++ src/php/Run/RunFacade.php
@@ -7,8 +7,8 @@ namespace Phel\Run;
 use Gacela\Framework\AbstractFacade;
 use Phel\Run\Application\Test\CpuCountDetector;
 use Phel\Run\Application\Test\ParallelTestOrchestrator;
-use Phel\Run\Domain\Repl\EvalResult;
 use Phel\Shared\CompileOptions;
+use Phel\Shared\Eval\EvalResult;
 use Phel\Shared\Exceptions\CompilerException;
 use Phel\Shared\Facade\RunFacadeInterface;
 use Phel\Shared\NamespaceInformation;
