+++ src/php/Shared/Eval/EvalError.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Phel\Run\Domain\Repl;
+namespace Phel\Shared\Eval;
 
 final readonly class EvalError
 {
