+++ tests/php/Unit/Nrepl/Domain/Op/EvalOpTest.php
@@ -8,9 +8,9 @@ use Phel\Nrepl\Application\Op\EvalOp;
 use Phel\Nrepl\Application\Op\EvalResultResponder;
 use Phel\Nrepl\Domain\Op\OpRequest;
 use Phel\Nrepl\Domain\Session\SessionRegistry;
-use Phel\Run\Domain\Repl\EvalError;
-use Phel\Run\Domain\Repl\EvalResult;
 use Phel\Shared\CompileOptions;
+use Phel\Shared\Eval\EvalError;
+use Phel\Shared\Eval\EvalResult;
 use Phel\Shared\Facade\RunFacadeInterface;
 use Phel\Shared\Printer\PrinterInterface;
 use PHPUnit\Framework\TestCase;
