+++ tests/php/Unit/Watch/Application/ReloadOrchestratorTest.php
@@ -7,8 +7,8 @@ namespace PhelTest\Unit\Watch\Application;
 use Gacela\Framework\Health\ModuleHealthCheckInterface;
 use Phel\Build\Domain\Compile\BuildOptions;
 use Phel\Build\Domain\Compile\CompiledFile;
-use Phel\Run\Domain\Repl\EvalResult;
 use Phel\Shared\CompileOptions;
+use Phel\Shared\Eval\EvalResult;
 use Phel\Shared\Exceptions\CompilerException;
 use Phel\Shared\Facade\BuildFacadeInterface;
 use Phel\Shared\Facade\RunFacadeInterface;
